Split the TAN command into two separate commands; now tempinfo and settemp have their own commands.
